backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.5.3
python-dotenv==1.0.0
email-validator==2.1.0
langchain==0.1.0
langchain-ollama==0.1.0
python-docx==1.1.0
markdown==3.5.1
weasyprint==60.1